+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.2] - 2019-04-15
 ### Added
 - `AVIF_NCLX_COLOUR_PRIMARIES_P3` (convenient mirrored value)
 - `avifNclxColourPrimariesFind()` - Finds a builtin avifNclxColourPrimaries and name by a set of primaries
@@ -33,5 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `avifVersion()` function
 
 [Unreleased]: https://github.com/joedrago/avif/compare/v0.1.1...HEAD
+[0.1.2]: https://github.com/joedrago/avif/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/joedrago/avif/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/joedrago/avif/releases/tag/v0.1.0